Sanitise: Move tests into new sub-directory
