+++ loki/transformations/sanitise/tests/test_associates.py
@@ -7,16 +7,13 @@
 
 import pytest
 
-from loki import (
-    BasicType, FindNodes, Subroutine, Module, fgen
-)
+from loki import BasicType, Subroutine
 from loki.frontend import available_frontends, OMNI
 from loki.ir import nodes as ir, FindNodes
 
 from loki.transformations.sanitise import (
     resolve_associates, merge_associates,
-    transform_sequence_association, ResolveAssociatesTransformer,
-    SanitiseTransformation
+    ResolveAssociatesTransformer,
 )
 
 
@@ -359,122 +356,3 @@ end subroutine merge_associates_simple
     call = FindNodes(ir.CallStatement).visit(routine.body)[0]
     b_c_n = call.kwarguments[0][1]  # b(c)%n
     assert b_c_n.scope == assocs[0]
-
-
-@pytest.mark.parametrize('frontend', available_frontends())
-def test_transform_sequence_assocaition_scalar_notation(frontend, tmp_path):
-    fcode = """
-module mod_a
-    implicit none
-
-    type type_b
-        integer :: c
-        integer :: d
-    end type type_b
-
-    type type_a
-        type(type_b) :: b
-    end type type_a
-
-contains
-
-    subroutine main()
-
-        type(type_a) :: a
-        integer :: k, m, n
-
-        real    :: array(10,10)
-
-        call sub_x(array(1, 1), 1)
-        call sub_x(array(2, 2), 2)
-        call sub_x(array(m, 1), k)
-        call sub_x(array(m-1, 1), k-1)
-        call sub_x(array(a%b%c, 1), a%b%d)
-
-    contains
-
-        subroutine sub_x(array, k)
-
-            integer, intent(in) :: k
-            real, intent(in)    :: array(k:n)
-
-        end subroutine sub_x
-
-    end subroutine main
-
-end module mod_a
-    """.strip()
-
-    module = Module.from_source(fcode, frontend=frontend, xmods=[tmp_path])
-    routine = module['main']
-
-    transform_sequence_association(routine)
-
-    calls = FindNodes(ir.CallStatement).visit(routine.body)
-
-    assert fgen(calls[0]).lower() == 'call sub_x(array(1:10, 1), 1)'
-    assert fgen(calls[1]).lower() == 'call sub_x(array(2:10, 2), 2)'
-    assert fgen(calls[2]).lower() == 'call sub_x(array(m:10, 1), k)'
-    assert fgen(calls[3]).lower() == 'call sub_x(array(m - 1:10, 1), k - 1)'
-    assert fgen(calls[4]).lower() == 'call sub_x(array(a%b%c:10, 1), a%b%d)'
-
-
-@pytest.mark.parametrize('frontend', available_frontends())
-@pytest.mark.parametrize('resolve_associate', [True, False])
-@pytest.mark.parametrize('resolve_sequence', [True, False])
-def test_transformation_sanitise(frontend, resolve_associate, resolve_sequence, tmp_path):
-    """
-    Test that the selective dispatch of the sanitisations works.
-    """
-
-    fcode = """
-module test_transformation_sanitise_mod
-  implicit none
-
-  type rick
-    real :: scalar
-  end type rick
-contains
-
-  subroutine test_transformation_sanitise(a, dave)
-    real, intent(inout) :: a(3)
-    type(rick), intent(inout) :: dave
-
-    associate(scalar => dave%scalar)
-      scalar = a(1) + a(2)
-
-      call vadd(a(1), 2.0, 3)
-    end associate
-
-  contains
-    subroutine vadd(x, y, n)
-      real, intent(inout) :: x(n)
-      real, intent(inout) :: y
-      integer, intent(in) :: n
-
-      x = x + 2.0
-    end subroutine vadd
-  end subroutine test_transformation_sanitise
-end module test_transformation_sanitise_mod
-"""
-    module = Module.from_source(fcode, frontend=frontend, xmods=[tmp_path])
-    routine = module['test_transformation_sanitise']
-
-    assoc = FindNodes(ir.Associate).visit(routine.body)
-    assert len(assoc) == 1
-    calls = FindNodes(ir.CallStatement).visit(routine.body)
-    assert len(calls) == 1
-    assert calls[0].arguments[0] == 'a(1)'
-
-    trafo = SanitiseTransformation(
-        resolve_associate_mappings=resolve_associate,
-        resolve_sequence_association=resolve_sequence,
-    )
-    trafo.apply(routine)
-
-    assoc = FindNodes(ir.Associate).visit(routine.body)
-    assert len(assoc) == 0 if resolve_associate else 1
-
-    calls = FindNodes(ir.CallStatement).visit(routine.body)
-    assert len(calls) == 1
-    assert calls[0].arguments[0] == 'a(1:3)' if resolve_sequence else 'a(1)'
